moving useFlags to auth
